Refactored Pet into separate animal classes
